+++ augur/mask.py
@@ -3,15 +3,15 @@ Mask specified sites from a VCF or FASTA file.
 """
 import os
 import sys
+from shlex import quote as shquote
 from shutil import copyfile
 
 from Bio.Seq import MutableSeq
 
 from .argparse_ import ExtendOverwriteDefault
 from .io.file import open_file
-from .io.sequences import read_sequences, write_sequences
+from .io.sequences import read_sequences, write_sequences, is_vcf
 from .io.shell_command_runner import run_shell_command
-from .io.vcf import shquote, is_vcf
 from .utils import load_mask_sites, VALID_NUCLEOTIDES
 
 def get_chrom_name(vcf_file):